fix(11): address code review findings CR-01, WR-01, WR-02
CR-01 (blocker): _resolve_placement_block did not guard against
non-list API responses — a 2xx response with a JSON dict or null
body caused AttributeError mid-sync. Add isinstance(blocks, list)
guard; returns None on malformed response.

WR-01: int(window_duration) in sync_gemini silently truncated float
durations and raised ValueError on '1.5'-style strings. Changed to
float(window_duration).

WR-02: insert_or_create_calendar_event used event.save() (full save),
risking overwrite of fields not managed by this function. Changed to
event.save(update_fields=list(fields.keys()) + ['modified']) to
update only the managed fields while still triggering the auto_now
modified timestamp.

+++ solsys_code/management/commands/sync_gemini_observation_calendar.py
@@ -121,7 +121,7 @@ class Command(BaseCommand):
                     start_dt = datetime.strptime(window_date, '%Y-%m-%d').replace(tzinfo=dt_timezone.utc)
                     time_dt = datetime.strptime(window_time_str, '%H:%M')
                     start_time = start_dt.replace(hour=time_dt.hour, minute=time_dt.minute)
-                    end_time = start_time + timedelta(hours=int(window_duration))
+                    end_time = start_time + timedelta(hours=float(window_duration))
                 else:
                     # GEM-WINDOW-02: fall back to ToO-type prefix when no explicit window is present.
                     # description_str is guaranteed non-None here (raw fallback took the explicit-window branch).
